ml-services/requirements.txt
fastapi 
uvicorn 
pydantic 
pyyaml 
openapi-spec-validator 
google-genai